Test: Isolate image module coverage
- add a TUIkitImage-only test target
- move decoder, URL loading, pixel, and converter tests
- keep public Image view checks in the umbrella suite
- remove the direct image dependency from integration tests
