+++ Tests/TUIkitImageTests/ImageTests.swift
@@ -5,7 +5,7 @@
 //  License: MIT
 
 import Testing
-@testable import TUIkit
+@testable import TUIkitImage
 
 // MARK: - RGBA Tests
 
@@ -284,34 +284,6 @@ struct ASCIIConverterTests {
 
         #expect(lines.count == 1)
         #expect(lines[0].contains("38;5;"))
-    }
-}
-
-// MARK: - Image View Tests
-
-@Suite("Image View Tests")
-@MainActor
-struct ImageViewTests {
-
-    @Test("Image initializes with file source")
-    func imageFileInit() {
-        let image = Image(.file("/path/to/image.png"))
-        #expect(image.source == .file("/path/to/image.png"))
-    }
-
-    @Test("Image initializes with URL source")
-    func imageURLInit() {
-        let image = Image(.url("https://example.com/image.png"))
-        #expect(image.source == .url("https://example.com/image.png"))
-    }
-
-    @Test("ImageSource equality works")
-    func imageSourceEquality() {
-        let source = ImageSource.file("/path/a.png")
-        let matchingSource = ImageSource.file("/path/a.png")
-        let differentSource = ImageSource.url("https://example.com")
-        #expect(source == matchingSource)
-        #expect(source != differentSource)
     }
 }
